+++ getic/CEditReplaceDlg.cpp
@@ -88,7 +88,7 @@ BOOL CEditReplaceDlg::OnInitDialog()
 BOOL CEditReplaceDlg::DoHighlightText()
 {
     ASSERT(m_pBuddy != NULL);
-    DWORD dwSearchFlags = 0;
+    size_t dwSearchFlags = 0;
     if (m_bMatchCase)
        dwSearchFlags |= FIND_MATCH_CASE;
     if (m_bWholeWord)
