+++ getic/Crc.h
@@ -6,11 +6,11 @@
 class CCrc  
 {
 public:
-    CCrc(DWORD poly =  0x04c11db7L);
+    CCrc(size_t poly =  0x04c11db7L);
 	virtual ~CCrc();
-    DWORD update_crc(DWORD crc_accum, BYTE *data_blk_ptr, int data_blk_size);
+    size_t update_crc(size_t crc_accum, BYTE *data_blk_ptr, int data_blk_size);
 
-    DWORD           polynomial;
+    size_t           polynomial;
 
     unsigned long   crc_table[256];
